Maps and Location Added the locate me button and my location marker.  These are tied to our custom location manager such that we are not double humping the battery.  I.E. the my location icon and button use locations only from our location manager, no new location manager is started.
Created a MAGE application to store and start the location services.

Added a 'follow me' option.  This will pan the map if your location marker moves outside of the maps bounds.  I.E. to save on battey I do not pan the map ever time we get a new location, that could be comsuming, I only pan if you move outside the visible bounds.  This pan will reset the map center to your location