Could not pick date on datepicker
The way the current data picker was setup it was hard to pick the date.  I changed the datapicker mode to
calendar mode.

I also updated the code to set the values of the date and time pickers to the value that is currently set on the
observation.  It seemed annoying that it you set a date/time and went back in to change it, the value was reset.