Re design how attachments are shown in observation feed I have been wanted to do this for a while.  Found a bug where the attachment view was not being sized correctly when a cell in the adapter was being recycled.  Instead of fixing this I removed that view that was taking up a portion of the right side of the observation.  The observation view now has a horizontal scroll layout for all attachments at the bottom of the observation.
Each attachment in that scroll view is clickable and will navigate to the image view activity.  This did
not happen before.  This makes it easier to view the attachment right from the news feed.

I also increased the size of the attachement thumbs in the scroll from 100x100 to 300x300 in both the
feed and the observation view.