Move disclaimer
Disclaimer info show/hide and title/text is now being pulled from the server.  If on the diclaimer now shows after successful auth.  Since this happens after successful auth we now keep track of if the user
has accepted the disclaimer.  This will prevent the user from backgrounding the app at the disclaimer activity and re-lauching mage to 'skip' accepting the disclaimer.

Reworked some of the server URL logic in the LoginActivity.  If the app is build with a blank url the server url lock will be unlocked and the server url text
will will have focus.

Also pulling /api everytime you hit the login activity.  This will help us verify, before login if the server/client versions are compatible.  It will also pull the dislaimer info that we need.

fixes #83